% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{annotate_uniprot}
\alias{annotate_uniprot}
\title{Annotate uniprot ids}
\usage{
annotate_uniprot(values, up = autonomics.annotate::connect_to_uniprot(values))
}
\arguments{
\item{values}{uniprot accessions (character vector)}

\item{up}{uniprot.ws connection}
}
\value{
data.table with uniprot annotations
}
\description{
Annotate uniprot ids
}
\examples{
\dontrun{
require(magrittr)
values <- c("A0A024R4M0", "M0R210", "G3HSF3", "P31941")
up <- values \%>\% autonomics.annotate::connect_to_uniprot()
autonomics.annotate::annotate_uniprot(values, up) \%>\% print()
}
}
