% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_uniprot.R
\name{annotate_uniprot_with_webservice}
\alias{annotate_uniprot_with_webservice}
\title{Annotate uniprot ids}
\usage{
annotate_uniprot_with_webservice(
  x,
  connection = connect_to_uniprot(x),
  columns = c("SUBCELLULAR-LOCATIONS", "KEGG", "GO-ID", "INTERPRO")
)
}
\arguments{
\item{x}{uniprot accessions (character vector)}

\item{connection}{uniprot.ws connection}

\item{columns}{character vector}
}
\value{
data.table with uniprot annotations
}
\description{
Annotate uniprot ids
}
\examples{
\dontrun{
    require(magrittr)
    x <- c('A0MZ66', 'A1KZ92', 'A1L020', 'A1XBS5', 'S4R350', 'A4D1P6')
    up <- x \%>\% connect_to_uniprot()
    annotate_uniprot_with_webservice(
       x, up, columns =  'SUBCELLULAR-LOCATIONS') \%>\% print()
}
}
