% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ANNOTATED_ORGANISMS}
\alias{ANNOTATED_ORGANISMS}
\title{Annotated organisms}
\format{
An object of class \code{list} of length 6.
}
\usage{
ANNOTATED_ORGANISMS
}
\description{
Annotated organisms
}
\keyword{datasets}
