% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_chebi.R
\name{chebi_to_smiles}
\alias{chebi_to_smiles}
\title{Get canonical smiles from CHEBI}
\usage{
chebi_to_smiles(CHEBIID)
}
\arguments{
\item{CHEBIID}{chEBI ID (numeric, character, or factor)}
}
\value{
canonical smiles string
}
\description{
Web scraping based, so non-vectorizable and slow.
No REST interface available (as per Dec 2017), only Perl and Java SOAP API.
Metabolon files do not give cheEBI annotations.
}
\examples{
   chebi_to_smiles('28733')
}
