% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_uniprot.R
\name{clean_existence}
\alias{clean_existence}
\title{Clean uniprot "protein existence" values}
\usage{
clean_existence(x)
}
\arguments{
\item{x}{uniprot protein existence values (character vector)}
}
\value{
cleaned values (character vector)
}
\description{
Clean uniprot "protein existence" values
}
\examples{
x <- c( 'Evidence at protein level',
        'Evidence at transcript level',
        'Inferred from homology',
        'Predicted',
        'Uncertain')
clean_existence(x)
}
