% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{clean_genes}
\alias{clean_genes}
\title{Clean uniprot "gene" values}
\usage{
clean_genes(x)
}
\arguments{
\item{x}{uniprot "gene" values (character vector)}
}
\value{
cleaned uniprot gene values (character vector)
}
\description{
Clean uniprot "gene" values
}
\examples{
\dontrun{
require(magrittr)
x <- c("A0A024R4M0", "M0R210", "G3HSF3")
up <- x \%>\% connect_to_uniprot()
annotations <- x \%>\% fetch_uniprot_annotations(up)
annotations$GENE \%>\% print()
annotations$GENE \%>\% clean_genes() \%>\% print()
}
}
