% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{clean_locations}
\alias{clean_locations}
\title{Clean uniprot location values}
\usage{
clean_locations(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Clean uniprot location values
}
\examples{
require(magrittr)
\dontrun{
    uniprot_vector <- c('A0MZ66', 'A1KZ92', 'A1L020', 'A1XBS5',
                        'S4R350', 'A4D1P6')
    up <- connect_to_uniprot(uniprot_vector)
    dt <- AnnotationDbi::select(
            up,
            keys    = uniprot_vector,
            columns = 'SUBCELLULAR-LOCATIONS') \%>\%
          data.table::data.table()
   x <- dt$`SUBCELLULAR-LOCATIONS`
   x \%>\% clean_locations()
}
}
