% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{clean_reviewed}
\alias{clean_reviewed}
\title{Clean uniprot "reviewed" values}
\usage{
clean_reviewed(x)
}
\arguments{
\item{x}{uniprot "reviewed" values (character vector)}
}
\value{
cleaned values (character vector)
}
\description{
Clean uniprot "reviewed" values
}
\examples{
\dontrun{
require(magrittr)
x <- c("A0A024R4M0", "M0R210", "G3HSF3")
up <- x \%>\% connect_to_uniprot()
annotations <- x \%>\% fetch_uniprot_annotations(up)
annotations$REVIEWED \%T>\% print() \%>\% clean_reviewed()
}
}
