% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{clean_score}
\alias{clean_score}
\title{Clean uniprot "score" values}
\usage{
clean_score(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vectors
}
\description{
Clean uniprot "score" values
}
\examples{
\dontrun{
    require(magrittr)
    x <- c("A0A024R4M0", "M0R210", "G3HSF3")
    up <- values \%>\% connect_to_uniprot()
    annotations <- x \%>\% fetch_uniprot_annotations(up)
    annotations$SCORE \%T>\% print() \%>\% clean_score()
}
}
