% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{clean_uniprot_existence_values}
\alias{clean_uniprot_existence_values}
\title{Clean uniprot "protein existence" values}
\usage{
clean_uniprot_existence_values(values)
}
\arguments{
\item{values}{uniprot protein existence values (character vector)}
}
\value{
cleaned values (character vector)
}
\description{
Clean uniprot "protein existence" values
}
\examples{
\dontrun{
require(magrittr)
values <- c("A0A024R4M0", "M0R210", "G3HSF3")
up <- values \%>\% autonomics.annotate::connect_to_uniprot()
annotations <- values \%>\% autonomics.annotate::fetch_uniprot_annotations(up)
annotations$EXISTENCE \%T>\% print() \%>\% autonomics.annotate::clean_uniprot_existence_values()
}
}
