% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{clean_uniprot_gene_values}
\alias{clean_uniprot_gene_values}
\title{Clean uniprot "gene" values}
\usage{
clean_uniprot_gene_values(values)
}
\arguments{
\item{values}{uniprot "gene" values (character vector)}
}
\value{
cleaned uniprot gene values (character vector)
}
\description{
Clean uniprot "gene" values
}
\examples{
\dontrun{
require(magrittr)
values <- c("A0A024R4M0", "M0R210", "G3HSF3")
up <- values \%>\% autonomics.annotate::connect_to_uniprot()
annotations <- values \%>\% autonomics.annotate::fetch_uniprot_annotations(up)
annotations$GENE \%>\% print()
annotations$GENE \%>\% autonomics.annotate::clean_uniprot_gene_values() \%>\% print()
}
}
