% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{clean_uniprot_proteinname_values}
\alias{clean_uniprot_proteinname_values}
\title{Clean uniprot "protein name" values}
\usage{
clean_uniprot_proteinname_values(values)
}
\arguments{
\item{values}{uniprot protein name values (character vector)}
}
\value{
cleaned uniprot protein name values (character vector)
}
\description{
Clean uniprot "protein name" values
}
\examples{
protein1 <- paste0("Rho GTPase-activating protein 10 ",
                   "(GTPase regulator associated with focal adhesion kinase 2) ",
                   "(Graf-related protein 2) ",
                   "(Rho-type GTPase-activating protein 10)")
protein2 <- paste0("Dystroglycan (Dystrophin-associated glycoprotein 1) ",
                   "[Cleaved into: Alpha-dystroglycan (Alpha-DG); Beta-dystroglycan (Beta-DG)]")
protein3 <- paste0("Bifunctional 3'-phosphoadenosine 5'-phosphosulfate synthase 1 ",
                   "(PAPS synthase 1) (PAPSS 1) (Sulfurylase kinase 1) (SK 1) (SK1) ",
                   "[Includes: Sulfate adenylyltransferase (EC 2.7.7.4) (ATP-sulfurylase) ",
                   "(Sulfate adenylate transferase) (SAT); Adenylyl-sulfate kinase (EC 2.7.1.25) ",
                   "(3'-phosphoadenosine-5'-phosphosulfate synthase) (APS kinase) ",
                   "(Adenosine-5'-phosphosulfate 3'-phosphotransferase) ",
                   "(Adenylylsulfate 3'-phosphotransferase)]")
values <- c(protein1, protein2, protein3)
clean_uniprot_proteinname_values(values)
}
