% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{connect_to_uniprot}
\alias{connect_to_uniprot}
\title{Open a uniprot.ws connection}
\usage{
connect_to_uniprot(values)
}
\arguments{
\item{values}{uniprot ids}
}
\value{
uniprot.ws connection
}
\description{
Infers organism from uniprot ids and opens a
uniprot webservice connection.
}
\examples{
\dontrun{
   require(magrittr)
   values <- c("A0A024R4M0", "M0R210", "G3HSF3")
   up <- values \%>\% autonomics.annotate::connect_to_uniprot()
}
}
