% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensg2entrezg.R
\name{ensg_to_entrezg}
\alias{ensg_to_entrezg}
\title{Map ensg to entrezg}
\usage{
ensg_to_entrezg(
  x,
  organism = infer_organism(keys = x, keytype = "ensg"),
  verbose = TRUE
)
}
\arguments{
\item{x}{ensg vector}

\item{organism}{any value in ANNOTATED_ORGANISMS}

\item{verbose}{logical}
}
\description{
Map ensg to entrezg
}
\examples{
   x <- c("ENSG00000066248", "ENSG00000066279", "ENSG99999", NA_character_)
   autonomics.annotate::ensg_to_entrezg(x)
}
