% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_FEATURE_IDENTIFIERS.R
\docType{data}
\name{FEATURE_ID_TYPES}
\alias{FEATURE_ID_TYPES}
\title{Feature identifer types}
\format{
An object of class \code{character} of length 2.
}
\usage{
FEATURE_ID_TYPES
}
\description{
Feature identifer types
}
\keyword{datasets}
