% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpro.R
\name{fetch_interpro_maps}
\alias{fetch_interpro_maps}
\title{Fetch interpro maps: interproid -> interproname}
\usage{
fetch_interpro_maps(verbose = FALSE)
}
\arguments{
\item{verbose}{\code{\link{logical}} rendering the function silent (or not)}
}
\value{
list (interproids_to_interpronames)
}
\description{
Fetch interpro maps: interproid -> interproname
}
\examples{
fetch_interpro_maps()
}
\author{
Johannes Graumann
}
