% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_kegg.R
\name{fetch_keggpathway_maps}
\alias{fetch_keggpathway_maps}
\title{Fetch keggpathway mapps}
\usage{
fetch_keggpathway_maps(kegggeneids, n_request_packet = 10, keep_raw = FALSE)
}
\arguments{
\item{kegggeneids}{\code{\link{character}} object representing KEGG gene/protein IDs,
for which information is retrieved via \code{\link[KEGGREST]{keggGet}} from package
\pkg{KEGGREST}}

\item{n_request_packet}{single \code{\link{numeric}} representing the number of accessions
sent to KEGG in a single request (web service restricted to 10)}

\item{keep_raw}{single \code{\link{logical}} indicating whether the complete
entire data returned by KEGG should be retained or
solely preprocessed products thereoff}
}
\value{
list (keggpathwayid_to_keggpathwayname, keggeneid_to_keggpathwayid)
}
\description{
kegggeneid -> keggpathwayid -> keggpathwayname
}
\examples{
kegggeneids <- c("dre:793465", "dre:571876", "dre:798771", "dre:564511")
fetch_keggpathway_maps(kegggeneids)
}
\author{
Johannes Graumann
}
