% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ontology_maps.R
\name{fetch_interpro_maps}
\alias{fetch_interpro_maps}
\alias{fetch_kegg_maps}
\title{Functions to generate maps for various ontology accession-based mapping operations}
\usage{
fetch_interpro_maps(
  interpro_url = paste0(c("http:/", "ftp.ebi.ac.uk", "pub", "databases", "interpro",
    "current", "names.dat"), collapse = "/"),
  verbose = FALSE
)

fetch_kegg_maps(kegg_geneids, n_request_packet = 10, keep_raw = FALSE)
}
\arguments{
\item{interpro_url}{single \code{\link{character}} representing the URL to the current
\code{names.dat} file from InterPro (default is \code{\link{paste0}}d to avoid line
length issues in the package documentation)}

\item{verbose}{\code{\link{logical}} rendering the function silent (or not)}

\item{kegg_geneids}{\code{\link{character}} object representing KEGG gene/protein IDs,
for which information is retrieved via \code{\link[KEGGREST]{keggGet}} from package
\pkg{KEGGREST}}

\item{n_request_packet}{single \code{\link{numeric}} representing the number of accessions
send to KEGG in a single request (web service restricted to 10)}

\item{keep_raw}{single \code{\link{logical}} indicating whether the complete
entire data returned by KEGG should be retained or solely preprocessed products thereoff}
}
\description{
Functions to generate maps for various ontology accession-based mapping operations
}
\examples{
library(magrittr)
# Fetch InterPro map
fetch_interpro_maps() \%>\%
  str()

# Fetch small KEGG map (gene/protein id specific, as whole data set not public)
c("dre:793465", "dre:571876", "dre:798771", "dre:564511") \%>\%
  fetch_kegg_maps() \%>\%
  str()
}
\author{
Johannes Graumann
}
