% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{fetch_uniprot_annotations}
\alias{fetch_uniprot_annotations}
\title{Fetch uniprot annotations}
\usage{
fetch_uniprot_annotations(values, up)
}
\arguments{
\item{values}{uniprot ids (character vector)}

\item{up}{uniprot.ws connection (output of autonomics.annotate::connect_to_uniprot)}
}
\value{
dataframe with annotations
}
\description{
Fetch uniprot annotations
}
\examples{
\dontrun{
require(magrittr)
values <- c("A0A024R4M0", "M0R210", "G3HSF3")
up <- values \%>\% autonomics.annotate::connect_to_uniprot()
values \%>\% autonomics.annotate::fetch_uniprot_annotations(up)
}
}
