% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdb.R
\name{gsymbol_to_entrezg}
\alias{gsymbol_to_entrezg}
\title{Map gsymbol to entrezg}
\usage{
gsymbol_to_entrezg(x, organism, verbose = TRUE)
}
\arguments{
\item{x}{gene symbol vector}

\item{organism}{any value in ANNOTATED_ORGANISMS}

\item{verbose}{logical}
}
\value{
unnamed character vector
}
\description{
Map gsymbol to entrezg
}
\examples{
   x <- c("NGEF", "ASPM", 'ABRACADABRA', NA_character_)
   organism <- 'Homo sapiens'
   gsymbol_to_entrezg(x, organism)
}
