% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_orgdb.R
\name{infer_organism}
\alias{infer_organism}
\title{Infer organism}
\usage{
infer_organism(keys, keytype, verbose = TRUE)
}
\arguments{
\item{keys}{keys}

\item{keytype}{any value in names(FEATURE_IDENTIFIERS)}

\item{verbose}{whether to report message (logical)}
}
\value{
organism name
}
\description{
Infer organism
}
\examples{
require(magrittr)
keys <- c('ENSG00000000003',  'ENSG00000000005', 'ENSG00000000419')
infer_organism(keys, 'ensg')
}
