% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetscan.R
\name{infer_organism_from_mirs}
\alias{infer_organism_from_mirs}
\title{Infer organism}
\usage{
infer_organism_from_mirs(mir)
}
\arguments{
\item{mir}{vector of micrRNAs}
}
\value{
name of infered organism (value in TARGETSCAN_ORGANISMS)
}
\description{
Infer organism
}
\examples{
require(magrittr)
mir <- c("hsa-mir-7-5p", "hsa-mir-29a-3p", "hsa-mir-199a-5p ")
mir \%>\% infer_organism_from_mirs()
mir <- 'rna-blablabla'
infer_organism_from_mirs(mir)
}
