% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_if_necessary.R
\name{install_package_if_necessary}
\alias{install_package_if_necessary}
\title{Install package if necessary}
\usage{
install_package_if_necessary(pkg)
}
\arguments{
\item{pkg}{package}
}
\description{
Install package if necessary
}
