% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpro.R
\name{interproids_to_interpronames}
\alias{interproids_to_interpronames}
\title{Convert interproid to interproname}
\usage{
interproids_to_interpronames(interproids,
  mapping_object = fetch_interpro_maps(), verbose = FALSE)
}
\arguments{
\item{interproids}{\code{\link{character}} with InterPro ids ('IPR*')}

\item{mapping_object}{mapping object as returned by \code{\link{fetch_interpro_maps}};}

\item{verbose}{\code{\link{logical}(1)}}
}
\value{
Named \code{\link{character}}: name = interproid, value = interproname
}
\description{
Convert interproid to interproname
}
\examples{
library(magrittr)

# Fetch interpromap
    interpromap  <- fetch_interpro_maps()

# Interproid -> Interproname
    interproids <-  c("IPR038348", "IPR011771", "IPRNotInDatabase")
    interproids \%>\% interproids_to_interpronames(interpromap)
    interproids \%>\% interproids_to_interpronames()

}
\author{
Johannes Graumann
}
