% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{kegggeneids_to_keggpathwayids}
\alias{kegggeneids_to_keggpathwayids}
\alias{keggpathwayids_to_keggpathwaynames}
\title{Map between kegggeneid, keggpathwayid and keggpathwayname}
\usage{
kegggeneids_to_keggpathwayids(kegggeneids,
  maps = fetch_keggpathway_maps(kegggeneids), verbose = FALSE)

keggpathwayids_to_keggpathwaynames(keggpathwayids,
  maps = fetch_keggpathway_maps(keggpathwayids), verbose = FALSE)
}
\arguments{
\item{kegggeneids}{\code{\link{character}} with kegggeneids}

\item{maps}{object returned by \code{\link{fetch_keggpathway_maps}};}

\item{verbose}{\code{\link{logical}(1)}}

\item{keggpathwayids}{\code{\link{character}} with keggpathwayids}
}
\value{
Named \code{\link{character}}: name = keggpathwayid, value = keggpathwayname
}
\description{
Map between kegggeneid, keggpathwayid and keggpathwayname
}
\examples{
library(magrittr)

# Fetch keggpathway maps
    kegggeneids <- c("dre:793465", "dre:571876", "dre:798771")
    kegg_map <- kegggeneids \%>\% fetch_keggpathway_maps()
    kegg_map \%>\% lapply(head, 1)

# kegggeneid -> keggpathwayid
    keggpathwayids <- kegggeneids_to_keggpathwayids(kegggeneids, kegg_map)
    keggpathwayids

## keggpathwayid -> keggpathwayname
    keggpathwayids \%>\%
    strsplit(split = ';')     \%>\%
    unlist(use.names = FALSE) \%>\%
    keggpathwayids_to_keggpathwaynames(kegg_map)
}
\author{
Johannes Graumann
}
