% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_orgdb.R
\name{load_orgdb}
\alias{load_orgdb}
\title{Load orgdb object}
\usage{
load_orgdb(organism)
}
\arguments{
\item{organism}{value in ANNOTATED_ORGANISMS}
}
\value{
AnnotationDbi::OrgDb package
}
\description{
Load orgdb object
}
\examples{
load_orgdb(organism = 'Homo sapiens')
load_orgdb(organism = 'Xenopus laevis')
}
