% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_targetscan.R
\name{load_targetscan}
\alias{load_targetscan}
\title{Load targetscan}
\usage{
load_targetscan(organism)
}
\arguments{
\item{organism}{Any value in TARGETSCAN_ORGANISMS}
}
\value{
data.table
}
\description{
Load targetscan
}
\examples{
\dontrun{
   load_targetscan('H.sapiens')
   load_targetscan('M.musculus')
   load_targetscan('D.rerio')
}
}
