% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_accession_mapping.R
\name{pathway_accessions_to_names}
\alias{pathway_accessions_to_names}
\alias{gene_accessions_to_pathway_accessions}
\alias{gene_accessions_to_pathway_accessions}
\title{Accession-based mapping/interconversion for multiple ontologies}
\usage{
pathway_accessions_to_names(accessions, ontology = c("auto", "interpro",
  "kegg"), mapping_object = NULL, ..., verbose = FALSE)

gene_accessions_to_pathway_accessions(accessions, ontology = c("auto",
  "kegg"), mapping_object = NULL, ..., verbose = FALSE)
}
\arguments{
\item{accessions}{\code{\link{character}} object representing InterPro accessions
('IPR*'), for which names are to be retrieved}

\item{ontology}{single \code{\link{character}} defining the ontology used or triggering
ontology-deduction (\code{auto})}

\item{mapping_object}{a mapping object as returned by e.g. \code{\link{fetch_interpro_maps}};
if \code{NULL} map fetching is handled internally}

\item{...}{further arguments handed on to map fetching}

\item{verbose}{single \code{\link{logical}} indicating whether retreival progress etc.
are printed}
}
\value{
Named \code{\link{character}} of InterPro names (using the corresponding
accessions as names)
}
\description{
Accession-based mapping/interconversion for multiple ontologies
}
\examples{
library(magrittr)
# InterPro: Simple call - determine ontology and fetch data on the fly
c("IPR038348", "IPR011771", "IPR031559", "IPR007277") \%>\%
  pathway_accessions_to_names()

# InterPro: Same thing - Format ERROR
\dontrun{
c("IPR038348", "IPR011771", "IPR031559", "IPR007277", "MyFavoriteAccession") \%>\%
  pathway_accessions_to_names()}

# InterPro: Formatting trickery to demonstrate empty string returns
c("IPR038348", "IPR011771", "IPRMyFavoriteAccession", "IPR007277") \%>\%
  pathway_accessions_to_names()

# InterPro: Pre-fetch map (reduced networking)
mo_ip <- fetch_interpro_maps()
c("IPR038348", "IPR011771", "IPR031559", "IPR007277") \%>\%
  pathway_accessions_to_names(mapping_object = mo_ip)

# A KEGG workflow - fetching here is gene id dependent
## A. Prefetch data for a bunch of gene/protein accessions
mo_kegg <- c("dre:793465", "dre:571876", "dre:798771", "dre:564511") \%>\%
  fetch_kegg_maps()
str(mo_kegg)
## B. Retrieve pathways for a specific gene/protein accession
(pwa <- gene_accessions_to_pathway_accessions('dre:798771', mapping_object = mo_kegg))
## C. Get names for retrieved pathways
pwa \%>\%
  strsplit(split = ';') \%>\%
  unlist(use.names = FALSE) \%>\%
  pathway_accessions_to_names(mapping_object = mo_kegg)
}
