% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{pubchem_to_smiles}
\alias{pubchem_to_smiles}
\title{Map a vector of PUBCHEM IDS to canonical smiles}
\usage{
pubchem_to_smiles(PUBCHEMIDS)
}
\arguments{
\item{PUBCHEMIDS}{character vector of PUBCHEM IDS}
}
\value{
character vector with canonical smiles
}
\description{
Map a vector of PUBCHEM IDS to canonical smiles
}
\examples{
require(magrittr)
PUBCHEMIDS <- c("11988421", "10236635", "5283147",  "91477",    "500",      "5283142")
PUBCHEMIDS \%>\% autonomics.annotate::pubchem_to_smiles()
}
