% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_uniprot.R
\name{read_uniprot_annotations}
\alias{read_uniprot_annotations}
\title{Read uniprot annotations from fastafile}
\usage{
read_uniprot_annotations(
  fastafile,
  fastafields = c("GENES", "PROTEIN-NAMES", "REVIEWED", "EXISTENCE", "ENTRYNAME",
    "ORGNAME", "VERSION")
)
}
\arguments{
\item{fastafile}{path to fasta file}

\item{fastafields}{character vector}
}
\value{
data.table (uniprot, genename, proteinname, reviewed, existence)
}
\description{
Read uniprot annotations from fastafile
}
\note{
EXISTENCE values are always those of canonical isoform
}
\examples{
require(magrittr)
if (require(autonomics.data)){
    fastafile <- system.file(
                     'extdata/fasta/uniprot_hsa_20140515.fasta',
                     package = 'autonomics.data')
    fastafile \%>\% read_uniprot_annotations()
}
}
