% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{has_cleaved}
\alias{has_cleaved}
\alias{rm_cleaved}
\title{rm "includes" constructs from protein name values}
\usage{
has_cleaved(x)

rm_cleaved(x)
}
\arguments{
\item{x}{uniprot protein name values}
}
\value{
updated values (include constructs removed)
}
\description{
rm "includes" constructs from protein name values
}
\examples{
require(magrittr)
values <- paste0(
    "Dystroglycan (Dystrophin-associated glycoprotein 1) ",
    "[Cleaved into: Alpha-dglycan (Alpha-DG); Beta-dystroglycan (Beta-DG)]")
values \%>\% has_cleaved()
values \%>\% rm_cleaved()
}
