% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{has_prefix_rs}
\alias{has_prefix_rs}
\alias{rm_prefix_rs}
\title{Detect or rm prefix (3R)-}
\usage{
has_prefix_rs(x)

rm_prefix_rs(x)
}
\arguments{
\item{x}{uniprot protein name values}
}
\description{
Detect or rm prefix (3R)-
}
\examples{
require(magrittr)
x <- paste0(
    "Very-long-chain (3R)-3-hydroxyacyl-CoA dehydratase 2 ",
    "(EC 4.2.1.134) ",
    "(3-hydroxyacyl-CoA dehydratase 2) ",
    "(HACD2) ",
    "(Protein-tyrosine phosphatase-like member B)")
x \%>\% has_prefix_rs()
x \%>\% rm_prefix_rs()
}
