% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_uniprot.R
\name{rm_protein_synonyms}
\alias{rm_protein_synonyms}
\title{Rm synonyms from uniprot proteinname values}
\usage{
rm_protein_synonyms(x)
}
\arguments{
\item{x}{character vector: uniprot proteinname values}
}
\value{
character vector with standard proteinname values
}
\description{
Rm synonyms from uniprot proteinname values
}
\examples{
require(magrittr)
x <- paste0(
    "Rho GTPase-activating protein 10 ",
    "(GTPase regulator associated with focal adhesion kinase 2) ",
    "(Graf-related protein 2) ",
    "(Rho-type GTPase-activating protein 10)")
x \%>\% rm_protein_synonyms()
}
