% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{annotate_proteingroups}
\alias{annotate_proteingroups}
\title{Annotate proteingroups}
\usage{
annotate_proteingroups(object, fastafile, fastafields = c("GENES",
  "PROTEIN-NAMES", "EXISTENCE", "REVIEWED"), wsconnection = NULL,
  wsfields = c("SUBCELLULAR-LOCATIONS", "INTERPRO"))
}
\arguments{
\item{object}{SummerizedExperiment with proteinGroups data}

\item{fastafile}{path to fastafile}

\item{webservice}{NULL or character vector (annotations to fetch from uniprot webservice)}
}
\value{
SummarizedExperiment with annotations
}
\description{
Annotate proteingroups
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::stemcomp.proteinratios[1:100, ]
   fastafile <- '../data/uniprot_hsapiens_20140515_68406entries.fasta'
   if (file.exists(fastafile)){
      object \%>\% autonomics.annotate::annotate_proteingroups(fastafile = fastafile)
   }
   \dontrun{
      wsconnection <- autonomics.annotate::connect_to_uniprot(c("A0AV96", "A0AVF1", "A0AVT1"))
   }
}
}
