% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{annotate_uniprot_with_webservice}
\alias{annotate_uniprot_with_webservice}
\title{Annotate uniprot ids}
\usage{
annotate_uniprot_with_webservice(values,
  up = autonomics.annotate::connect_to_uniprot(values),
  columns = c("SUBCELLULAR-LOCATIONS", "KEGG", "GO-ID", "INTERPRO"))
}
\arguments{
\item{values}{uniprot accessions (character vector)}

\item{up}{uniprot.ws connection}
}
\value{
data.table with uniprot annotations
}
\description{
Annotate uniprot ids
}
\examples{
\dontrun{
require(magrittr)
values <- c('A0MZ66', 'A1KZ92', 'A1L020', 'A1XBS5', 'S4R350', 'A4D1P6')
up <- values \%>\% autonomics.annotate::connect_to_uniprot()
autonomics.annotate::annotate_uniprot_with_webservice(
   values, up, columns =  'SUBCELLULAR-LOCATIONS') \%>\% print()
}
}
