% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{clean_existence}
\alias{clean_existence}
\title{Clean uniprot "protein existence" values}
\usage{
clean_existence(x)
}
\arguments{
\item{x}{uniprot protein existence values (character vector)}
}
\value{
cleaned values (character vector)
}
\description{
Clean uniprot "protein existence" values
}
\examples{
\dontrun{
require(magrittr)
    x <- c("A0A024R4M0", "M0R210", "G3HSF3")
    up <- x \%>\% connect_to_uniprot()
    annotations <- x \%>\% fetch_uniprot_annotations(up,'EXISTENCE')
    annotations$EXISTENCE \%T>\% print() \%>\% clean_existence()
}
}
