% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_uniprot.R
\name{clean_genes}
\alias{clean_genes}
\title{Clean uniprot "gene" values}
\usage{
clean_genes(x)
}
\arguments{
\item{x}{uniprot "gene" values (character vector)}
}
\value{
cleaned uniprot gene values (character vector)
}
\description{
Clean uniprot "gene" values
}
\examples{
x <- c('RPS9 hCG_2009111', 'RPS16 hCG_43353', NA_character_)
x \%>\% clean_genes()
}
