% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_uniprot.R
\name{clean_reviewed}
\alias{clean_reviewed}
\title{Clean uniprot "reviewed" values}
\usage{
clean_reviewed(x)
}
\arguments{
\item{x}{uniprot "reviewed" values (character vector)}
}
\value{
cleaned values (character vector)
}
\description{
Clean uniprot "reviewed" values
}
\examples{
x <- c('unreviewed', 'reviewed', NA_character_)
clean_reviewed(x)
}
