% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_uniprot.R
\name{clean_score}
\alias{clean_score}
\title{Clean uniprot "score" values}
\usage{
clean_score(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vectors
}
\description{
Clean uniprot "score" values
}
\examples{
x <- c("1 out of 5", "2 out of 5", NA_character_)
clean_score(x)
}
