% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{fetch_uniprot_annotations}
\alias{fetch_uniprot_annotations}
\title{Fetch uniprot annotations}
\usage{
fetch_uniprot_annotations(x, up, columns = c("SUBCELLULAR-LOCATIONS",
  "KEGG", "GO-ID", "INTERPRO"))
}
\arguments{
\item{x}{uniprot ids (character vector)}

\item{up}{uniprot.ws connection (output of connect_to_uniprot)}

\item{columns}{character vector}
}
\value{
dataframe with annotations
}
\description{
Fetch uniprot annotations
}
\examples{
\dontrun{
    require(magrittr)
    x <- c("A0A024R4M0", "M0R210", "G3HSF3")
    up <- x \%>\% connect_to_uniprot()
    x \%>\% fetch_uniprot_annotations(up)
}
}
