% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orgdb_name.R
\name{get_orgdb_name}
\alias{get_orgdb_name}
\title{Get orgdb name}
\usage{
get_orgdb_name(organism)
}
\arguments{
\item{organism}{any value in \code{names(ANNOTATED_ORGANISMS)}}
}
\description{
Get orgdb name
}
\examples{
get_orgdb_name('Homo sapiens')
}
