% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_install_package_if_necessary.R
\name{install_package_if_necessary}
\alias{install_package_if_necessary}
\title{Install package if necessary}
\usage{
install_package_if_necessary(pkg)
}
\arguments{
\item{pkg}{character(1): R package name}
}
\value{
R package name
}
\description{
Install package if necessary
}
\examples{
\dontrun{
    install_package_if_necessary('pracma')
}
}
