% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_orgdb.R
\name{load_go_genes}
\alias{load_go_genes}
\title{Load GO genes}
\usage{
load_go_genes(goid, organism, gene = "ENTREZID")
}
\arguments{
\item{goid}{GO ID}

\item{organism}{'Homo sapiens'}

\item{gene}{'ENTREZID', 'ENSEMBL', 'UNIPROT', 'SYMBOL', 'GENENAME'}
}
\value{
character vector
}
\description{
Load GO genes
}
\examples{
require(magrittr)
load_go_genes(goid = 'GO:0006954', organism = 'Homo sapiens', gene = 'ENTREZID') \%>\% head()
load_go_genes(goid = 'GO:0006954', organism = 'Homo sapiens', gene = 'ENSEMBL')  \%>\% head()
load_go_genes(goid = 'GO:0006954', organism = 'Homo sapiens', gene = 'UNIPROT') \%>\% head()
load_go_genes(goid = 'GO:0006954', organism = 'Homo sapiens', gene = 'SYMBOL') \%>\% head()
load_go_genes(goid = 'GO:0006954', organism = 'Homo sapiens', gene = 'GENENAME') \%>\% head()
}
