% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{load_uniprot_annotations_from_fastafile}
\alias{load_uniprot_annotations_from_fastafile}
\title{Load uniprot annotations from protein sequence fastafile}
\usage{
load_uniprot_annotations_from_fastafile(fastafile,
  fastafields = c("CANONICAL", "GENES", "PROTEIN-NAMES", "REVIEWED",
  "EXISTENCE", "ENTRYNAME", "ORGID", "ORGNAME", "VERSION"))
}
\arguments{
\item{fastafile}{path to fasta file}

\item{fastafields}{character vector}
}
\value{
data.table with fields: (uniprot, genename, proteinname, reviewed, existence)
}
\description{
Load uniprot annotations from protein sequence fastafile
}
\note{
EXISTENCE values are always those of the canonical isoform (no isoform-level resolution for this field)
}
\examples{
require(magrittr)
fastafile <- '../data/uniprot_hsapiens_20140515_68406entries.fasta'
if (file.exists(fastafile)){
   fastafile \%>\% autonomics.annotate::load_uniprot_annotations_from_fastafile()
}
}
