% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_targetscan.R
\name{mir_to}
\alias{mir_to}
\title{Map mir to genes}
\usage{
mir_to(mir, to)
}
\arguments{
\item{mir}{vector of microRNAs}

\item{to}{'ensg' or 'gsymbol'}
}
\value{
named list that maps each mir to its target genes.
}
\description{
Map vector of microRNAs to target genes
}
\examples{
require(magrittr)
mir <- c('hsa-mir-7-5p', 'hsa-mir-29a-3p', 'hsa-mir-199a-5p ')
mir \%>\% mir_to('gsymbol')
}
