% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{rm_protein_synonyms}
\alias{rm_protein_synonyms}
\title{Rm synonyms from uniprot proteinname values}
\usage{
rm_protein_synonyms(values)
}
\arguments{
\item{values}{uniprot proteinname values}
}
\value{
character vector with standard proteinname values
}
\description{
Rm synonyms from uniprot proteinname values
}
\examples{
require(magrittr)
values <- paste0("Rho GTPase-activating protein 10 ",
                 "(GTPase regulator associated with focal adhesion kinase 2) ",
                 "(Graf-related protein 2) ",
                 "(Rho-type GTPase-activating protein 10)")
values \%>\% rm_protein_synonyms()
}
